refactor(tui): split the story row layout into wrap-plan and gutter modules
