+++ tui/src/screens/story/row-layout.ts
@@ -1,5 +1,4 @@
 import { chapterWord, extentLabel } from "../../chapter-model.js";
-import type { KeyAction } from "../../keys.js";
 import type {
   ChapterDividerRow,
   ChapterSummaryRow,
@@ -7,17 +6,7 @@ import type {
   StoryRow,
   StoryViewModel
 } from "../../model.js";
-import type { StoryScreenState, StreamView } from "../../state.js";
-import {
-  MAX_HUMAN_EDIT_RANGES,
-  MAX_REWRITTEN_SPANS,
-  type StoryNode,
-  type TextRange
-} from "../../../../shared/types.js";
-import {
-  STARTER_LOGO_LINES,
-  STARTER_LOGO_TEXT
-} from "../../../../shared/starter-vault.js";
+import type { StoryScreenState } from "../../state.js";
 import { takeStrip } from "./density.js";
 import {
   fitLine,
@@ -29,59 +18,33 @@ import {
   type FrameLine,
   type FrameSegment
 } from "./frame.js";
-import { wrapText, type ProseStyle, type StyleRun, type WrappedLine, type WrapCache } from "../../wrap.js";
+import { wrapText, type ProseStyle, type WrappedLine, type WrapCache } from "../../wrap.js";
 import {
   registerNextDeadline,
   type FrameDeadlineCollector
 } from "../../animation-deadline.js";
-import { streamForPart, streamTrimBounds } from "../../stream-text.js";
-import type { WrapContentIdentity } from "../../wrap.js";
+import { streamForPart } from "../../stream-text.js";
 import { STORY_GUTTER } from "../../composer-geometry.js";
 import { thoughtGutterContext, type ThoughtGutterContext } from "../../reasoning-model.js";
+import { unfoldedThoughtBlock } from "./thought-block.js";
 import {
-  focusedFoldedThoughtLine,
-  ghostThoughtMark,
-  thinkingGutterLine0,
-  thinkingGutterLine1,
-  unfoldedThoughtBlock
-} from "./thought-block.js";
-
-/** The complete braille cycle. The dots travel once around the cell and arrive
- *  back where they started, so the mark reads as one turn. Four of these ten
- *  marks carry the dots a quarter of the way around and then snap back to the
- *  top, which reads as a stall rather than as progress. */
-const STREAM_LIVENESS_MARKS = [
-  "⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"
-] as const;
-const STREAM_LIVENESS_FRAME_MS = 250;
+  starterLogoSourceStart,
+  storyPartWrapIdentity,
+  wrapIdentityContext,
+  wrapPart
+} from "./wrap-plan.js";
+import {
+  actionHint,
+  gutterFor,
+  gutterRowsFor,
+  streamLivenessMark
+} from "./gutter.js";
 
 export interface StoryRowLayout {
   height: number;
   stickyGutter: StickyStoryGutter | null;
   stickyPrompt: StickyStoryPrompt | null;
   render(): FrameLine[];
-}
-
-export interface StoryPartWrapPlan {
-  partId: string;
-  width: number;
-  text: string;
-  runs: readonly StyleRun<ProseStyle>[];
-  sourceStart: number;
-  compactLogo: boolean;
-  identity: WrapContentIdentity;
-  stream: StreamView | null;
-  appending: boolean;
-  appendStart: number | null;
-}
-
-export interface StoryPartWrapInput {
-  id: string;
-  node: Pick<StoryNode, "text">;
-  isSummary: boolean;
-  humanSpans: readonly TextRange[];
-  /** Ranges of prose a rewrite replaced (issue #319) — see `humanSpans`. */
-  rewrittenSpans: readonly TextRange[];
 }
 
 export interface StickyStoryGutter {
@@ -130,8 +93,12 @@ export function layoutStoryRow(
   const stream = streamForPart(state.stream, row.id);
   const streaming = stream !== null;
   const thought = thoughtGutterContext(row, rowIndex, state, streaming, stream);
-  const prefixMask = (narrow ? 1 : 0) | (state.showInstructions ? 2 : 0) | (row.isSummary ? 4 : 0)
-    | (thought.show && thought.unfolded ? 8 : 0);
+  const prefixFlags: PartPrefixFlags = {
+    boundary: narrow,
+    prompt: state.showInstructions,
+    summaryHeader: row.isSummary,
+    thought: thought.kind === "shown" && !thought.folded
+  };
   const identityContext = wrapIdentityContext(row, state);
   const identity = storyPartWrapIdentity(
     row,
@@ -152,30 +119,32 @@ export function layoutStoryRow(
   const wrappedBody = cache.lineCount(row.id, wrapWidth, wrappedTextInput, identity)
     ?? prepare().wrapped.length;
   const wrapped = wrappedBody + (sourceStart === 0 ? 0 : 1);
-  const gutterRows = streaming && !narrow ? 2
-    : focused && !narrow && !row.isSummary
-      ? (row.siblingCount > 1 ? 2 : 1) + GUTTER_VERBS.length + (thought.show && !thought.unfolded ? 1 : 0)
-      : 0;
+  // Built once per row: `gutterRows.length` (below) and every per-line lookup
+  // in `renderPartBody` both read this same list, so the row's gutter height
+  // and its content can never drift apart the way two hand-derived formulas
+  // could. Only computed when it will actually be used — the streaming
+  // gutter's fixed 2-line pair (`gutterFor`, gutter.ts) takes priority over
+  // this whenever both are true.
+  const gutterRows = focused && !narrow && !streaming && !row.isSummary
+    ? gutterRowsFor(row, thought)
+    : null;
+  const gutterRowCount = streaming && !narrow ? 2 : gutterRows?.length ?? 0;
   let prefix: FrameLine[] | null = null;
   const preparePrefix = () => prefix ??= partPrefix(
-    row, rowIndex, allParts, state, measure, narrow, focused, streaming, prefixMask, thought, deadlines
+    row, rowIndex, allParts, state, measure, narrow, focused, streaming, prefixFlags, thought, deadlines
   );
-  const expandedPrompt = (prefixMask & 2) !== 0 && state.expandedPromptIds.has(row.id);
+  const expandedPrompt = prefixFlags.prompt && state.expandedPromptIds.has(row.id);
   // An unfolded thought block's height is as variable as an expanded
   // prompt's own wrap — neither fits the fixed `prefixHeight` formula, so
   // both fall back to measuring the real, fully-built prefix once.
-  const needsMeasuredPrefix = expandedPrompt || (prefixMask & 8) !== 0;
-  const prefixRows = needsMeasuredPrefix ? preparePrefix().length : prefixHeight(prefixMask);
+  const needsMeasuredPrefix = expandedPrompt || prefixFlags.thought;
+  const prefixRows = needsMeasuredPrefix ? preparePrefix().length : prefixHeight(prefixFlags);
   const stickyGutter = focused && !narrow && !row.isSummary && !streaming
-    ? {
-        start: prefixRows,
-        lines: Array.from({ length: gutterRows }, (_, lineIndex) =>
-          gutterFor(row, true, false, lineIndex, thought, state.now, deadlines))
-      }
+    ? { start: prefixRows, lines: gutterRows ?? [] }
     : null;
   const promptRowIndex = narrow ? 1 : 0;
-  const partRows = prefixRows + Math.max(wrapped, gutterRows);
-  const stickyPrompt: StickyStoryPrompt | null = (prefixMask & 2) !== 0 && !expandedPrompt
+  const partRows = prefixRows + Math.max(wrapped, gutterRowCount);
+  const stickyPrompt: StickyStoryPrompt | null = prefixFlags.prompt && !expandedPrompt
     ? {
         start: promptRowIndex,
         partRows,
@@ -188,7 +157,7 @@ export function layoutStoryRow(
     stickyPrompt,
     render: () => [
       ...preparePrefix(),
-      ...renderPartBody(row, state, focused, narrow, prepare(), gutterRows, thought, deadlines)
+      ...renderPartBody(row, state, focused, narrow, prepare(), gutterRowCount, gutterRows, thought, deadlines)
     ]
   };
 }
@@ -247,6 +216,16 @@ function renderChapterSummary(
   return lines;
 }
 
+/** The four independent rows a part's prefix can carry, named instead of
+ *  packed into a bitmask — see `prefixHeight` for why `thought` is excluded
+ *  from the sum every other field contributes to. */
+interface PartPrefixFlags {
+  readonly boundary: boolean;
+  readonly prompt: boolean;
+  readonly summaryHeader: boolean;
+  readonly thought: boolean;
+}
+
 function partPrefix(
   part: StoryPart,
   rowIndex: number,
@@ -256,26 +235,16 @@ function partPrefix(
   narrow: boolean,
   focused: boolean,
   streaming: boolean,
-  mask: number,
+  prefix: PartPrefixFlags,
   thought: ThoughtGutterContext,
   deadlines?: FrameDeadlineCollector
 ): FrameLine[] {
   const lines: FrameLine[] = [];
-  if ((mask & 1) !== 0) lines.push(renderBoundary(part, measure, focused, streaming, state.now, deadlines));
-  if ((mask & 2) !== 0) {
+  if (prefix.boundary) lines.push(renderBoundary(part, measure, focused, streaming, state.now, deadlines));
+  if (prefix.prompt) {
     lines.push(...renderPrompt(part, rowIndex, state.expandedPromptIds.has(part.id), measure, narrow));
   }
-  if ((mask & 8) !== 0) {
-    // Above the prose, the same slot the expanded instruction renders in —
-    // see `renderPrompt` just above. Ordered after the prompt so reading
-    // order stays "what was asked" → "what the model thought" → "what it
-    // wrote". `resolved` is never null here: `thought.show` is what set this
-    // bit, and `thoughtGutterContext` only ever computes it alongside one.
-    for (const row of unfoldedThoughtBlock(thought.hit, measure, thought.resolved!)) {
-      lines.push(prefixLine(narrow, row.gutter, row.prose));
-    }
-  }
-  if ((mask & 4) !== 0) {
+  if (prefix.summaryHeader) {
     const previousSummary = allParts.slice(0, part.number - 1).findLastIndex((candidate) => candidate.isSummary);
     const start = previousSummary + 2;
     const end = part.number - 1;
@@ -283,11 +252,27 @@ function partPrefix(
     const header = `§ summary · parts ${start}–${end} · ${rawWords.toLocaleString("en-US")} → ${part.stub.words} words · replaces the text above`;
     lines.push(prefixLine(narrow, narrow ? [] : [segment("◈", "summary")], [segment(truncate(header, measure), "summary")]));
   }
+  // Above the prose, the same slot the expanded instruction renders in — see
+  // `renderPrompt` above. Ordered after the summary header so reading order
+  // stays "what the row is" → "what was asked" → "what the model thought" →
+  // "what it wrote" — a summary take names itself before it says what led to
+  // it. `prefix.thought` narrows `thought.kind` to `"shown"` by construction
+  // (`layoutStoryRow` derives both from the same check), but the type system
+  // has no way to know that, so the `kind` check below is the proof.
+  if (prefix.thought && thought.kind === "shown") {
+    for (const row of unfoldedThoughtBlock(thought.hit, measure, thought.resolved)) {
+      lines.push(prefixLine(narrow, row.gutter, row.prose));
+    }
+  }
   return lines;
 }
 
-function prefixHeight(mask: number): number {
-  return Number((mask & 1) !== 0) + Number((mask & 2) !== 0) + Number((mask & 4) !== 0);
+/** Sum of the prefix's fixed-height rows. `thought` is deliberately excluded:
+ *  an unfolded thought block wraps to as many rows as its text needs, so its
+ *  height is measured once from the real prefix (`needsMeasuredPrefix`,
+ *  `layoutStoryRow`) rather than counted here. */
+function prefixHeight(prefix: PartPrefixFlags): number {
+  return Number(prefix.boundary) + Number(prefix.prompt) + Number(prefix.summaryHeader);
 }
 
 function renderPrompt(
@@ -348,7 +333,8 @@ function renderPartBody(
   focused: boolean,
   narrow: boolean,
   prepared: ReturnType<typeof wrapPart>,
-  gutterRows: number,
+  gutterRowCount: number,
+  gutterRows: readonly FrameLine[] | null,
   thought: ThoughtGutterContext,
   deadlines?: FrameDeadlineCollector
 ): FrameLine[] {
@@ -362,19 +348,19 @@ function renderPartBody(
   if (compactLogo) {
     lines.push(prefixLine(
       narrow,
-      gutterFor(part, focused, streaming, 0, thought, state.now, deadlines),
+      gutterFor(part, streaming, 0, thought, gutterRows, state.now, deadlines),
       compactStarterLogo(focused)
     ));
   }
   for (const line of wrapped) {
     const lineIndex = lines.length;
-    lines.push(prefixLine(narrow, gutterFor(part, focused, streaming, lineIndex, thought, state.now, deadlines),
+    lines.push(prefixLine(narrow, gutterFor(part, streaming, lineIndex, thought, gutterRows, state.now, deadlines),
       styledWrapped(line, focused ? "prose" : "prose · dim", focused ? "human edit" : "human edit dim",
         "rewritten", state, part, streaming && !appending, deadlines, sourceStart)));
   }
   const proseTip = lines.length - 1;
-  for (let lineIndex = lines.length; lineIndex < gutterRows; lineIndex += 1) {
-    lines.push(prefixLine(narrow, gutterFor(part, focused, streaming, lineIndex, thought, state.now, deadlines), []));
+  for (let lineIndex = lines.length; lineIndex < gutterRowCount; lineIndex += 1) {
+    lines.push(prefixLine(narrow, gutterFor(part, streaming, lineIndex, thought, gutterRows, state.now, deadlines), []));
   }
   // Machine insertion stays visually distinct from the writer's solid block.
   if (streaming && proseTip >= 0) lines[proseTip]!.push(segment("▏", "chrome"));
@@ -382,290 +368,10 @@ function renderPartBody(
   // shows only the dim caret, on the row the `thinking` gutter line claims —
   // `wrapped` is empty (nothing to loop above), so nothing has planted the
   // caret yet the way the `proseTip` branch above just did for ordinary text.
-  else if (streaming && thought.thinking && lines.length > 0) lines[0]!.push(segment("▏", "chrome"));
+  else if (streaming && thought.kind === "shown" && thought.thinking && lines.length > 0) {
+    lines[0]!.push(segment("▏", "chrome"));
+  }
   return lines;
-}
-
-function wrapPart(
-  part: StoryPart,
-  stream: StreamView | null,
-  measure: number,
-  cache: WrapCache<ProseStyle>,
-  identityContext: { source: object; settledLength: number },
-  sourceStart?: number
-) {
-  const plan = storyPartWrapPlan(
-    part,
-    stream,
-    measure,
-    identityContext.settledLength,
-    identityContext.source,
-    stream,
-    sourceStart
-  );
-  return {
-    stream: plan.stream,
-    appending: plan.appending,
-    sourceStart: plan.sourceStart,
-    compactLogo: plan.compactLogo,
-    wrapped: cache.wrap(plan.partId, plan.width, plan.text, plan.runs, plan.identity)
-  };
-}
-
-/** Canonical prose-wrap input shared by synchronous rows and cold prewarming. */
-export function storyPartWrapPlan(
-  part: StoryPartWrapInput,
-  stream: StreamView | null,
-  measure: number,
-  settledLength = part.node.text.length,
-  identitySource: object = part.node,
-  streamIdentity: object | null = stream,
-  projectedSourceStart?: number
-): StoryPartWrapPlan {
-  const appending = stream?.append === true;
-  const sourceText = part.node.text;
-  const width = measure - (part.isSummary ? 2 : 0);
-  const sourceStart = projectedSourceStart
-    ?? (part.isSummary ? 0 : starterLogoSourceStart(sourceText, part.humanSpans, width));
-  const text = sourceStart === 0 ? sourceText : sourceText.slice(sourceStart);
-  // Wrapping aligns style boundaries while it already owns grapheme
-  // segmentation. Keeping this raw avoids a second uninterruptible scan before
-  // resumable cold work begins.
-  const sourceStreamingStart = appending && sourceText.length > settledLength
-    ? Math.max(0, settledLength)
-    : null;
-  const streamingStart = sourceStreamingStart === null
-    ? null
-    : Math.max(0, sourceStreamingStart - sourceStart);
-  const logoRuns: StyleRun<ProseStyle>[] = sourceStart === 0
-    ? starterLogoRuns(text, part.humanSpans)
-    : [];
-  const humanRuns = provenanceRuns(part.humanSpans, MAX_HUMAN_EDIT_RANGES, sourceStart, text.length, "human");
-  // A rewritten span (issue #319) marks prose the model wrote over the
-  // writer's own words — a weaker claim than a human span, which means the
-  // writer touched the passage since. A human edit over a rewritten span
-  // already reclaims it server-side (`rewrittenSpansAfterHumanEdit`,
-  // shared/human-edit.ts), so the two should rarely overlap by the time a
-  // plan reaches here — but `resolveProvenanceOverlay` below resolves any
-  // overlap and orders the result the same way regardless, rather than
-  // trusting that argument. The wrap engine paints every style run it is
-  // given, in the order given, without sorting or blending
-  // (materialize-runs, wrap.ts; clipRuns, wrap.ts), so an unresolved overlap
-  // or an out-of-order run would draw the same characters twice.
-  const rewrittenRuns = provenanceRuns(part.rewrittenSpans, MAX_REWRITTEN_SPANS, sourceStart, text.length, "rewritten");
-  let runs = resolveProvenanceOverlay(logoRuns, humanRuns, rewrittenRuns);
-  if (streamingStart !== null) {
-    // Every provenance run must end at or before the streaming boundary for
-    // the streaming run to stay last and the whole list to stay ascending.
-    // Truncate the merged list against that boundary here, once, instead of
-    // arguing separately that each family (in particular the starter logo,
-    // never otherwise clipped to it) already stays inside it. This step
-    // alone only removes the overlap with [streamingStart, text.length) —
-    // it does not by itself guarantee every survivor ends at or before
-    // text.length, since a run could originally have reached past it. That
-    // guarantee is `provenanceRuns`'s clamp, above (Fix 2, issue #319
-    // review): without it, a run's tail past text.length survived this
-    // subtraction as a piece sitting *after* [streamingStart, text.length)
-    // in source order but pushed into the array *before* the streaming run
-    // below — ascending within the family, non-ascending overall, the same
-    // shape Fix 1 exists to rule out.
-    runs = subtractAscending(runs, [{ start: streamingStart, end: text.length }]);
-    runs.push({ start: streamingStart, end: text.length, style: "streaming" });
-  }
-  return {
-    partId: part.id,
-    width,
-    text,
-    runs,
-    sourceStart,
-    compactLogo: sourceStart > 0,
-    identity: storyPartWrapIdentity(
-      part,
-      stream,
-      settledLength,
-      identitySource,
-      streamIdentity
-    ),
-    stream,
-    appending,
-    appendStart: streamingStart
-  };
-}
-
-/** Shift and clip one span family (human or rewritten) into runs over the
- *  already-sliced wrap text, in the caller's declared style. `sourceStart`
- *  and the cap match the semantics `storyPartWrapPlan` already applied
- *  before this split existed; the streaming boundary is no longer clipped
- *  here — `resolveProvenanceOverlay`'s caller clips the merged result once,
- *  uniformly, including the starter logo this function never sees.
- *
- *  `textLength` additionally clamps every run's end to the wrap text itself
- *  (Fix 2, issue #319 review). Nothing here can prove a persisted span never
- *  names an offset past the node's current text — `validateNodeRewrittenSpans`
- *  and `validateVersionAttributions` (server/story-format.ts) reject that on
- *  encode and decode, but this view has no way to prove it only ever sees
- *  payloads those checks passed, and re-arguing that on every call is the
- *  "by argument" gap the rest of this module deliberately avoids. Clamping
- *  here, unconditionally, is what makes it hold by construction instead: an
- *  unclamped run surviving past `text.length` produced a stray piece *after*
- *  the streaming-boundary subtraction below, landing before the streaming
- *  run pushed next — non-ascending, the exact shape Fix 1 exists to rule
- *  out. */
-function provenanceRuns(
-  spans: readonly TextRange[],
-  cap: number,
-  sourceStart: number,
-  textLength: number,
-  style: ProseStyle
-): StyleRun<ProseStyle>[] {
-  const count = Math.min(spans.length, cap);
-  const runs: StyleRun<ProseStyle>[] = [];
-  for (let index = 0; index < count; index += 1) {
-    const range = spans[index]!;
-    const start = Math.max(range.start, sourceStart) - sourceStart;
-    const end = Math.min(range.end - sourceStart, textLength);
-    if (end > start) runs.push({ start, end, style });
-  }
-  return runs;
-}
-
-/** Merge the starter-logo, human, and rewritten run families into one
- *  ascending, disjoint list — replacing the three hand-ordered
- *  concatenations this diff found (logo first, then every human span, then
- *  every rewritten span), which is ascending within each family but not
- *  overall: a rewritten span earlier in the text landed after a human span
- *  starting later, and the renderer (`styledWrapped`, below) walks runs with
- *  a single monotonic cursor, so the run out of order re-sliced text its
- *  cursor had already passed and drew it twice. A human span wins any
- *  overlap with a rewritten span; both win over the merely decorative
- *  starter logo. Overlap should be rare in practice (see the call site's
- *  comment), but resolving it here, unconditionally, is what makes the
- *  invariant hold by construction rather than by that argument — including
- *  for the starter logo, whose own pristine check does not itself rule out
- *  a rewritten span landing inside it.
- *
- *  `humanRuns` and `rewrittenRuns` each arrive ascending and pairwise
- *  disjoint within their own family — guaranteed at the parse boundary
- *  (`parseRewrittenSpans`, `parseVersionAttributions`, server/story-format-
- *  facts.ts), which reject unsorted, overlapping, or non-positive ranges on
- *  load. That is exactly what lets `subtractAscending` and `mergeAscending`
- *  below settle every overlap and every ordering with one linear pass each,
- *  in place of a per-run call into the general-purpose `subtractRanges`
- *  (shared/human-edit.ts) plus a final sort: on this render-path call, once
- *  per visible row per frame, that per-run approach cost O(runs × cuts)
- *  allocations — 2.17 ms for one part at the 256-human/256-rewritten cap,
- *  roughly 65 ms across 30 visible rows — against 0.008 ms for the merge
- *  below on the same input (Fix 1, issue #319 review). */
-function resolveProvenanceOverlay(
-  logoRuns: readonly StyleRun<ProseStyle>[],
-  humanRuns: readonly StyleRun<ProseStyle>[],
-  rewrittenRuns: readonly StyleRun<ProseStyle>[]
-): StyleRun<ProseStyle>[] {
-  const rewritten = subtractAscending(rewrittenRuns, humanRuns);
-  const human = mergeAscending(humanRuns, rewritten);
-  const logo = subtractAscending(logoRuns, human);
-  return mergeAscending(human, logo);
-}
-
-/** Subtract an ascending, pairwise-disjoint list of `cuts` from an
- *  ascending, pairwise-disjoint list of `runs`, keeping each survivor's own
- *  style, in one linear pass over both instead of a nested one. `cutIndex`
- *  only ever advances, across the whole call, because both lists are
- *  ascending: no run can need a cut that an earlier run had already fully
- *  passed. A single-range `cuts` list (the streaming-boundary call below)
- *  trivially satisfies "ascending and disjoint" too, so this one function
- *  serves both that call and `resolveProvenanceOverlay` above. */
-function subtractAscending(
-  runs: readonly StyleRun<ProseStyle>[],
-  cuts: readonly TextRange[]
-): StyleRun<ProseStyle>[] {
-  const result: StyleRun<ProseStyle>[] = [];
-  let cutIndex = 0;
-  for (const run of runs) {
-    let start = run.start;
-    while (start < run.end) {
-      while (cutIndex < cuts.length && cuts[cutIndex]!.end <= start) cutIndex += 1;
-      const cut = cuts[cutIndex];
-      if (cut === undefined || cut.start >= run.end) {
-        result.push({ start, end: run.end, style: run.style });
-        break;
-      }
-      if (cut.start > start) result.push({ start, end: cut.start, style: run.style });
-      start = Math.max(start, cut.end);
-    }
-  }
-  return result;
-}
-
-/** Interleave two ascending, pairwise-disjoint run lists into one ascending
- *  list — the merge step of a merge sort, safe to use without the sort
- *  itself because both inputs already arrive ordered. Replaces the
- *  `[...a, ...b].sort(...)` `resolveProvenanceOverlay` used to close with,
- *  which re-sorted an already-mostly-ordered list on every call regardless
- *  of whether either family was even present. */
-function mergeAscending(
-  left: readonly StyleRun<ProseStyle>[],
-  right: readonly StyleRun<ProseStyle>[]
-): StyleRun<ProseStyle>[] {
-  const merged: StyleRun<ProseStyle>[] = [];
-  let leftIndex = 0;
-  let rightIndex = 0;
-  while (leftIndex < left.length && rightIndex < right.length) {
-    merged.push(left[leftIndex]!.start <= right[rightIndex]!.start
-      ? left[leftIndex++]!
-      : right[rightIndex++]!);
-  }
-  while (leftIndex < left.length) merged.push(left[leftIndex++]!);
-  while (rightIndex < right.length) merged.push(right[rightIndex++]!);
-  return merged;
-}
-
-const STARTER_LOGO_STYLES = [
-  "logo red",
-  "logo orange",
-  "logo yellow",
-  "logo green",
-  "logo cyan",
-  "logo blue",
-  "logo violet"
-] as const satisfies readonly ProseStyle[];
-
-const STARTER_LOGO_PREFIX = `${STARTER_LOGO_TEXT}\n\n`;
-const STARTER_LOGO_WIDTH = Math.max(...STARTER_LOGO_LINES.map(visibleWidth));
-
-function hasPristineStarterLogo(text: string, humanSpans: readonly TextRange[]): boolean {
-  return text.startsWith(STARTER_LOGO_PREFIX)
-    && !humanSpans.some((range) => range.start < STARTER_LOGO_TEXT.length);
-}
-
-function starterLogoSourceStart(
-  text: string,
-  humanSpans: readonly TextRange[],
-  measure: number
-): number {
-  return measure < STARTER_LOGO_WIDTH && hasPristineStarterLogo(text, humanSpans)
-    ? STARTER_LOGO_PREFIX.length
-    : 0;
-}
-
-/** Match the complete prefix. An edited or partial logo becomes ordinary
- * prose, which keeps a user change authoritative and avoids hidden markup. */
-function starterLogoRuns(
-  text: string,
-  humanSpans: readonly TextRange[]
-): StyleRun<ProseStyle>[] {
-  if (!hasPristineStarterLogo(text, humanSpans)) return [];
-  const runs: StyleRun<ProseStyle>[] = [];
-  let lineStart = 0;
-  for (const line of STARTER_LOGO_LINES) {
-    for (let band = 0; band < STARTER_LOGO_STYLES.length; band += 1) {
-      const start = lineStart + band * 3;
-      const end = Math.min(start + 3, lineStart + line.length);
-      if (start < end) runs.push({ start, end, style: STARTER_LOGO_STYLES[band]! });
-    }
-    lineStart += line.length + 1;
-  }
-  return runs;
 }
 
 function compactStarterLogo(focused: boolean): FrameLine {
@@ -673,208 +379,6 @@ function compactStarterLogo(focused: boolean): FrameLine {
     ? ["logo red", "logo yellow", "logo green", "logo violet"] as const
     : ["prose · dim", "prose · dim", "prose · dim", "prose · dim"] as const;
   return [..."1667"].map((digit, index) => segment(digit, roles[index]!));
-}
-
-/** O(1) content proof for immutable payload prose and one append-only stream.
- * The raw text remains on the plan for cold work, never for warm comparison. */
-export function storyPartWrapIdentity(
-  part: StoryPartWrapInput,
-  stream: StreamView | null,
-  settledLength: number,
-  source: object,
-  streamIdentity: object | null
-): WrapContentIdentity {
-  const appending = stream?.append === true;
-  const streamingStart = appending && part.node.text.length > settledLength
-    ? Math.max(0, settledLength)
-    : null;
-  const contentStream = stream === null
-    || streamIdentity === null
-    || appending && streamingStart === null
-    ? null
-    : streamIdentity;
-  const bounds = contentStream === null
-    ? { start: 0, end: 0 }
-    : appending
-      ? { start: 0, end: stream!.text.length }
-      : streamTrimBounds(stream!);
-  return {
-    source,
-    stream: contentStream,
-    streamStart: bounds.start,
-    streamEnd: bounds.end,
-    textLength: part.node.text.length
-  };
-}
-
-function wrapIdentityContext(
-  part: StoryPart,
-  state: StoryScreenState
-): { source: object; settledLength: number } {
-  const authoritative = state.payload.path[part.pathIndex];
-  const source = authoritative?.id === part.id ? authoritative : state.payload;
-  return {
-    source,
-    settledLength: authoritative?.id === part.id
-      ? authoritative.text.length
-      : part.node.text.length
-  };
-}
-
-/** Every thought-gutter decision `gutterFor`, `partPrefix` and
- *  `renderPartBody` need for one row, computed once in `layoutStoryRow` and
- *  threaded through rather than recomputed at each call site — the same
- *  reason `streaming`/`prepared` are computed once and passed down.
- *
- * `show`/`unfolded` are both false whenever `state.reasoning === "off"`:
- * `off` renders nothing, ever, and gating it once here is what makes every
- * downstream reader — the gutter marks, the unfolded block, the streaming
- * line — inherit that instead of each needing its own `!== "off"` check. */
-export interface GutterVerb { token: string | null; label: string; action: KeyAction }
-
-export const GUTTER_VERBS: ReadonlyArray<readonly GutterVerb[]> = [
-  [{ token: "␠", label: "continue", action: "continue" }, { token: "↵", label: "direct", action: "compose" }],
-  [{ token: "r", label: "retake", action: "regenerate" }, { token: "R", label: "reprompt", action: "retake-with-prompt" }],
-  [{ token: "w", label: "write", action: "write" }, { token: "e", label: "edit", action: "edit" }]
-];
-
-const GUTTER_VERB_COLUMN_GAP = 2;
-
-function gutterVerbLabel(verb: GutterVerb): string {
-  return verb.token === null ? verb.label : `${verb.token} ${verb.label}`;
-}
-
-/** Widest label in each menu column. Derived from the verbs themselves so a
- * relabelled verb re-measures instead of drifting out of its column. */
-const GUTTER_VERB_COLUMNS: readonly number[] = GUTTER_VERBS.reduce<number[]>((columns, row) => {
-  row.forEach((verb, index) => {
-    columns[index] = Math.max(columns[index] ?? 0, visibleWidth(gutterVerbLabel(verb)));
-  });
-  return columns;
-}, []);
-
-/** Every verb row is padded to this one width. The gutter is right-aligned, so
- * a uniform width is what gives the menu a single left edge and a shared second
- * column instead of a different indent per row. */
-const GUTTER_VERB_WIDTH = GUTTER_VERB_COLUMNS.reduce(
-  (total, column, index) => total + column + (index === 0 ? 0 : GUTTER_VERB_COLUMN_GAP),
-  0
-);
-
-function gutterVerbSegments(row: readonly GutterVerb[]): FrameLine {
-  const line = row.flatMap((verb, index): FrameLine => {
-    const label = gutterVerbLabel(verb);
-    const pad = (GUTTER_VERB_COLUMNS[index] ?? 0) - visibleWidth(label);
-    return [
-      ...(index === 0 ? [] : [segment(" ".repeat(GUTTER_VERB_COLUMN_GAP))]),
-      actionHint(label, verb.action),
-      ...(pad > 0 ? [segment(" ".repeat(pad))] : [])
-    ];
-  });
-  // Squaring the row off to the block width is what makes the right-aligned
-  // gutter give every verb row the same left edge.
-  return fitLine(line, GUTTER_VERB_WIDTH);
-}
-
-function gutterFor(
-  part: StoryPart,
-  focused: boolean,
-  streaming: boolean,
-  lineIndex: number,
-  thought: ThoughtGutterContext,
-  now = 0,
-  deadlines?: FrameDeadlineCollector
-): FrameLine {
-  if (streaming) {
-    // Reasoning is arriving and no prose has started: swap the usual
-    // writing/esc-stops pair for thinking/esc-peeks. Once prose starts, or
-    // once `reasoning` is off, this falls through to the ordinary pair —
-    // `thought.thinking` is already false in both cases (see
-    // `thoughtGutterContext`).
-    if (thought.thinking) {
-      if (lineIndex === 0) return thinkingGutterLine0(thought.resolved!.tokenCount);
-      if (lineIndex === 1) return thinkingGutterLine1(thought.hit);
-      return [];
-    }
-    if (lineIndex === 0) return [segment(`${streamLivenessMark(now, deadlines)} writing`, "focus / accent")];
-    if (lineIndex === 1) return [actionHint("esc stops", "cancel")];
-    return [];
-  }
-  if (focused) {
-    if (lineIndex === 0 && part.siblingCount > 1) {
-      return [segment(`¶ ${part.number} `, "chrome"), ...takeCounterSegments(part)];
-    }
-    if (lineIndex === 0) return part.isSummary
-      ? [segment(`¶ ${part.number} `, "chrome"), segment("◈", "summary")]
-      : [segment(`¶ ${part.number}`, "chrome")];
-    if (part.siblingCount > 1 && lineIndex === 1) {
-      const strip = takeStrip(part.takeIndex, part.siblingCount, part.takeSubtakes);
-      const segments = stripSegments(strip, part.takeIndex);
-      // Pad the strip out to the counter above it so the right-aligned gutter
-      // starts both on the same column: the dots read as belonging to the
-      // counter rather than as a third indent.
-      const width = segments.reduce((sum, item) => sum + visibleWidth(item.text), 0);
-      const pad = takeCounterWidth(part) - width;
-      return pad > 0 ? [...segments, segment(" ".repeat(pad))] : segments;
-    }
-    // Folded and focused: one extra row, right after the counter/dots and
-    // ahead of the verb menu — `layoutStoryRow`'s `gutterRows` and the verb
-    // offset just below both already account for it. Unfolded skips this:
-    // its own `thought · N`/`T folds` pair lives in the block's own prefix
-    // rows instead (`unfoldedThoughtBlock`), so repeating it here would say
-    // the same thing twice.
-    const thoughtLineIndex = part.siblingCount > 1 ? 2 : 1;
-    if (thought.show && !thought.unfolded && lineIndex === thoughtLineIndex) {
-      return focusedFoldedThoughtLine(thought.hit);
-    }
-    if (!part.isSummary) {
-      const verbOffset = (part.siblingCount > 1 ? 2 : 1) + (thought.show && !thought.unfolded ? 1 : 0);
-      const verbs = GUTTER_VERBS[lineIndex - verbOffset];
-      if (verbs !== undefined) return gutterVerbSegments(verbs);
-    }
-  }
-  if (lineIndex === 0) {
-    if (part.isSummary) return [segment("◈", "summary")];
-    if (part.siblingCount > 1) return [segment(`×${part.siblingCount}`, "chrome")];
-    // No fork tick or summary diamond claims this row's one unfocused gutter
-    // cell: give the ghost thought word the cell instead. A part that has
-    // both a fork count and a thought keeps the fork count — the reader's
-    // place in the tree outranks a decorative margin word, and focusing the
-    // part still reveals `thought · T shows` a moment later.
-    if (thought.show && !thought.unfolded) return ghostThoughtMark(thought.hit);
-  }
-  return [];
-}
-
-function takeCounterSegments(part: StoryPart): FrameLine {
-  return [
-    actionHint("‹", "take-previous", "focus / accent"),
-    segment(` take ${part.takeIndex}/${part.siblingCount} `, "focus / accent"),
-    actionHint("›", "take-next", "focus / accent")
-  ];
-}
-
-function takeCounterWidth(part: StoryPart): number {
-  return takeCounterSegments(part).reduce((sum, item) => sum + visibleWidth(item.text), 0);
-}
-
-function stripSegments(strip: ReturnType<typeof takeStrip>, currentTake: number): FrameLine {
-  if (strip.density === "gauge") {
-    return [
-      segment(strip.text.slice(0, strip.currentOffset), "chrome"),
-      segment(strip.text[strip.currentOffset] ?? "", "focus / accent",
-        { kind: "story-take", take: currentTake }),
-      segment(strip.text.slice(strip.currentOffset + 1), "chrome")
-    ];
-  }
-  const spacing = strip.density === "spaced" ? " " : "";
-  // The strip owns the glyphs so the ring rule lives in one place; this only
-  // colours them and hangs a click target on each take.
-  return strip.cells.map((glyph, index): FrameLine => [
-    ...(index === 0 ? [] : [segment(spacing, "chrome")]),
-    segment(glyph, index + 1 === currentTake ? "focus / accent" : "chrome",
-      { kind: "story-take", take: index + 1 })
-  ]).flat();
 }
 
 /** Repaint only the fixed-width gutter; preserve cell-accurate prose segments. */
@@ -920,16 +424,6 @@ function renderBoundary(
   const prefix = `── ${marker} `;
   return [segment("  "), segment(prefix, "chrome"),
     segment("─".repeat(Math.max(0, measure - visibleWidth(prefix))), "chrome")];
-}
-
-/** One indicator owns both stream labels and their next visible frame. */
-export function streamLivenessMark(
-  now: number,
-  deadlines?: FrameDeadlineCollector
-): string {
-  const frame = Math.floor(now / STREAM_LIVENESS_FRAME_MS);
-  deadlines?.at((frame + 1) * STREAM_LIVENESS_FRAME_MS);
-  return STREAM_LIVENESS_MARKS[frame % STREAM_LIVENESS_MARKS.length]!;
 }
 
 function prefixLine(narrow: boolean, gutter: FrameLine, prose: FrameLine): FrameLine {
@@ -1030,8 +524,4 @@ function centeredRule(label: string, measure: number): string {
   const remaining = Math.max(2, measure - visibleWidth(body));
   const left = Math.floor(remaining / 2);
   return `${"─".repeat(left)}${body}${"─".repeat(remaining - left)}`;
-}
-
-function actionHint(text: string, action: KeyAction, role: DisplayRole = "chrome"): FrameSegment {
-  return segment(text, role, { kind: "inline-action", action });
 }
